+++ packages/icheck-ts/src/cli.ts
@@ -1,22 +1,47 @@
 import process from "node:process";
-import { TinyCliCommand, tinyCliMain } from "@hiogawa/tiny-cli";
-import { version as packageVersion } from "../package.json";
-import { runCommand, runnerArgs } from "./runner";
+import { TinyCliCommand, arg, tinyCliMain } from "@hiogawa/tiny-cli";
+import {
+  name as packageName,
+  version as packageVersion,
+} from "../package.json";
+import { type ExportUsage, run } from "./runner";
 
 const command = new TinyCliCommand(
   {
     program: "icheck-ts",
     version: packageVersion,
     description: "Lint import and export usages",
-    args: runnerArgs,
+    args: {
+      files: arg.stringArray("Typescript files to lint"),
+      cache: arg.boolean("Enable caching"),
+      cacheLocation: arg.string("Cache directory location", {
+        default: `node_modules/.cache/${packageName}/v${packageVersion}`,
+      }),
+      ignore: arg.string("RegExp pattern to ignore export names", {
+        optional: true,
+      }),
+    },
   },
   async ({ args }) => {
-    const { result, unused } = await runCommand(args);
+    const result = run(args.files, { cache: args.cache });
 
-    if (args.debug) {
-      console.log(new Map(result.exportUsages));
+    // apply extra unused rules
+    const ignoreRegExp = args.ignore && new RegExp(args.ignore);
+    const ignoreComment = "icheck-ignore";
+
+    function isUsedExport(e: ExportUsage): boolean {
+      return Boolean(
+        e.used ||
+          (ignoreRegExp && e.name.match(ignoreRegExp)) ||
+          e.comment.includes(ignoreComment)
+      );
     }
 
+    const unused = [...result.exportUsages]
+      .map(([k, vs]) => [k, vs.filter((v) => !isUsedExport(v))] as const)
+      .filter(([_k, vs]) => vs.length > 0);
+
+    // parse error
     if (result.errors.length > 0) {
       console.log("* Parse errors");
       for (const { file, error } of result.errors) {
@@ -25,6 +50,7 @@ const command = new TinyCliCommand(
       process.exitCode = 1;
     }
 
+    // unused exports error
     if (unused.length > 0) {
       console.log("* Unused exports");
       for (const [file, entries] of unused) {
